revert to original structure
