--- lightning_hpo/loggers/__init__.py
@@ -1,2 +1,0 @@
-from lightning_hpo.loggers.config import BaseConfig, Loggers
-from lightning_hpo.loggers.wandb import Wandb
